detection_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ultralytics==8.3.0
pillow>=10.0.0
python-multipart>=0.0.6
